requirements.txt
flask==3.1.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.0
librosa==0.10.1
soundfile==0.12.1
tensorflow==2.15.0
pydub==0.25.1
matplotlib==3.8.4
scipy==1.12.0
audioread==3.0.1
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2
openai-whisper==20231117
torch==2.2.0
torchaudio==2.2.0
pytest==8.0.0
black==24.1.1
flake8==7.0.0
python-magic==0.4.27
ffmpeg-python==0.2.0
seaborn==0.13.2
jupyter==1.0.0
ipykernel==6.29.0
gunicorn==21.2.0 